+++ python/tvm/__init__.py
@@ -57,9 +57,6 @@ from . import testing
 # tvm.driver
 from .driver import build, lower
 
-# tvm.hybrid
-from . import hybrid
-
 # others
 from . import arith
