+++ python/tvm/relay/op/nn/_nn.py
@@ -22,11 +22,11 @@ import topi
 from topi.util import get_const_tuple
 
 from tvm.runtime import convert
+from tvm.te.hybrid import script
 from .. import op as reg
 from .. import strategy
 from ..op import OpPattern
 from .._tensor import elemwise_shape_func
-from ....hybrid import script
 
 # relu
 reg.register_broadcast_schedule("nn.relu")
